+++ packages/plugin-packages/mashroom-portal/src/backend/utils/resource-utils.ts
@@ -2,7 +2,8 @@
 import http from 'http';
 import https from 'https';
 import {pipeline} from 'stream/promises';
-import {resourceUtils, ResourceNotFoundError, ResourceFetchAbortedError} from '@mashroom/mashroom-utils';
+import {URL} from 'url';
+import {resourceUtils, ResourceNotFoundError, ResourceFetchAbortedError, httpAgentStatsUtils} from '@mashroom/mashroom-utils';
 import context from '../context/global-portal-context';
 
 import type {Agent as HttpAgent} from 'http';
@@ -10,14 +11,31 @@ import type {Agent as HttpsAgent} from 'https';
 import type {Response} from 'express';
 import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
 
+export type RequestMetrics = {
+    httpRequestCountTotal: number;
+    httpRequestTargetCount: Record<string, number>;
+    httpConnectionErrorCountTotal: number;
+    httpConnectionErrorTargetCount: Record<string, number>;
+    httpTimeoutCountTotal: number;
+    httpTimeoutTargetCount: Record<string, number>;
+}
+
 let httpAgent: HttpAgent | undefined;
 let httpsAgent: HttpsAgent | undefined;
+const requestMetrics: RequestMetrics = {
+    httpRequestCountTotal: 0,
+    httpRequestTargetCount: {},
+    httpConnectionErrorCountTotal: 0,
+    httpConnectionErrorTargetCount: {},
+    httpTimeoutCountTotal: 0,
+    httpTimeoutTargetCount: {},
+};
 
 export const isNotFoundError = (error: Error) => {
   return error instanceof ResourceNotFoundError;
 };
 
-export const isAbortedError = (error: Error) => {
+export const isTimeoutError = (error: Error) => {
     return error instanceof ResourceFetchAbortedError;
 };
 
@@ -36,6 +54,7 @@ export const setupResourceFetchHttpAgents = (logger: MashroomLogger) => {
 };
 
 export const streamResource = async (resourceUri: string, res: Response, logger: MashroomLogger): Promise<void> => {
+    countRemoteResourceRequest(resourceUri);
     const {fetchTimeoutMs} = context.portalPluginConfig.resourceFetchConfig;
     const abortController = new AbortController();
     const abortTimeout = setTimeout(() => {
@@ -61,13 +80,20 @@ export const streamResource = async (resourceUri: string, res: Response, logger:
             res.type(fileName);
         }
         await pipeline(stream, res);
-
+    } catch (e: any) {
+        if (isTimeoutError(e)) {
+             countRemoteResourceTimeout(resourceUri);
+        } else {
+            countRemoteResourceConnectionError(resourceUri);
+        }
+        throw e;
     } finally {
         clearTimeout(abortTimeout);
     }
 };
 
 export const getResourceAsString = async (resourceUri: string, logger: MashroomLogger): Promise<string> => {
+    countRemoteResourceRequest(resourceUri);
     const {fetchTimeoutMs} = context.portalPluginConfig.resourceFetchConfig;
     const abortController = new AbortController();
     const abortTimeout = setTimeout(() => {
@@ -80,7 +106,68 @@ export const getResourceAsString = async (resourceUri: string, logger: MashroomL
             httpAgent,
             httpsAgent,
         });
+    } catch (e: any) {
+        if (isTimeoutError(e)) {
+            countRemoteResourceTimeout(resourceUri);
+        } else {
+            countRemoteResourceConnectionError(resourceUri);
+        }
+        throw e;
     } finally {
         clearTimeout(abortTimeout);
     }
+};
+
+// Metrics
+
+const getProtocolAndHost = (uri: string) => {
+    const {protocol, host} = new URL(uri);
+    return `${protocol}//${host}`;
+};
+
+const countRemoteResourceRequest = (resourceUri: string) => {
+    if (resourceUri.startsWith('http')) {
+        const target = getProtocolAndHost(resourceUri);
+        requestMetrics.httpRequestCountTotal++;
+        if (!requestMetrics.httpRequestTargetCount[target]) {
+            requestMetrics.httpRequestTargetCount[target] = 0;
+        }
+        requestMetrics.httpRequestTargetCount[target]++;
+    }
+};
+
+const countRemoteResourceConnectionError = (resourceUri: string) => {
+    if (resourceUri.startsWith('http')) {
+        const target = getProtocolAndHost(resourceUri);
+        requestMetrics.httpConnectionErrorCountTotal++;
+        if (!requestMetrics.httpConnectionErrorTargetCount[target]) {
+            requestMetrics.httpConnectionErrorTargetCount[target] = 0;
+        }
+        requestMetrics.httpConnectionErrorTargetCount[target]++;
+    }
+};
+
+const countRemoteResourceTimeout = (resourceUri: string) => {
+    if (resourceUri.startsWith('http')) {
+        const target = getProtocolAndHost(resourceUri);
+        requestMetrics.httpTimeoutCountTotal++;
+        if (!requestMetrics.httpTimeoutTargetCount[target]) {
+            requestMetrics.httpTimeoutTargetCount[target] = 0;
+        }
+        requestMetrics.httpTimeoutTargetCount[target]++;
+    }
+};
+
+export const getRequestMetrics = () => requestMetrics;
+export const getHttpAgentMetrics = (logger: MashroomLogger) => {
+    if (httpAgent) {
+        return httpAgentStatsUtils.getAgentStats(httpAgent, logger);
+    }
+    return null;
+};
+export const getHttpsAgentMetrics = (logger: MashroomLogger) => {
+    if (httpsAgent) {
+        return httpAgentStatsUtils.getAgentStats(httpsAgent, logger);
+    }
+    return null;
 };