+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * HTTP Portal: Added metrics for remote resource requests (*mashroom_portal_remote_resources_*), like request count, error count and pool stats
  * HTTP Proxy: The proxies do no longer automatically add *x-forwarded-* headers, because if you are using public APIs you might not want
    to disclose details of your internal network. Added a new config property *createForwardedForHeaders* to explicitly turn this feature on again. 
    Also fixed the implementation, it takes now existing headers from reverse proxies into consideration and keeps them or